# ARROW-6131: [C++] Optimize the Arrow Non-Ascii-string-validation