Algorithm and code come from: https://github.com/cyb70289/utf8 (MIT LICENSE)

The patch would introduce a fast non-ASCII validation method with this new algorithm into Apache Arrow and give give a option to non-ascii case validation.

### Arm64 platform
Benchmark

Origin:
```
----------------------------------------------------------------
Benchmark                         Time           CPU Iterations
----------------------------------------------------------------
ValidateTinyAscii                 7 ns          7 ns  107435339   1.42978GB/s
ValidateTinyNonAscii             16 ns         16 ns   42655054   639.503MB/s
ValidateSmallAscii               29 ns         29 ns   24516945    4.4671GB/s
ValidateSmallAlmostAscii         91 ns         91 ns    7677848   1.51182GB/s
ValidateSmallNonAscii           175 ns        175 ns    4009837    731.98MB/s
ValidateLargeAscii            18821 ns      18814 ns      37194   4.95077GB/s
ValidateLargeAlmostAscii      64056 ns      64025 ns      10929   1.45533GB/s
ValidateLargeNonAscii        130321 ns     130249 ns       5375   732.909MB/s
```

The new algorithm:
```
----------------------------------------------------------------
Benchmark                         Time           CPU Iterations
----------------------------------------------------------------
ValidateTinyAscii                 6 ns          6 ns  116427650   1.59527GB/s
ValidateTinyNonAscii             17 ns         17 ns   41897276   628.046MB/s
ValidateSmallAscii              117 ns        117 ns    5964896   1113.14MB/s
ValidateSmallAlmostAscii        145 ns        145 ns    4819232    971.76MB/s
ValidateSmallNonAscii           118 ns        118 ns    5947924   1085.68MB/s
ValidateLargeAscii            82297 ns      82247 ns       8511   1.13246GB/s
ValidateLargeAlmostAscii      81145 ns      81138 ns       8627   1.14838GB/s
ValidateLargeNonAscii         81221 ns      81202 ns       8621   1.14805GB/s
```



### x86 platform
Benchmark

Origin:
```
----------------------------------------------------------------
Benchmark                         Time           CPU Iterations
----------------------------------------------------------------
ValidateTinyAscii                 3 ns          3 ns  180870752   3.02228GB/s
ValidateTinyNonAscii             10 ns         10 ns   73066734    1061.9MB/s
ValidateSmallAscii               11 ns         11 ns   61844060   11.2604GB/s
ValidateSmallAlmostAscii         53 ns         53 ns   10000000   2.60296GB/s
ValidateSmallNonAscii            88 ns         88 ns    7948330   1.41807GB/s
ValidateLargeAscii             4887 ns       4886 ns     140027   19.0622GB/s
ValidateLargeAlmostAscii      31479 ns      31474 ns      22286    2.9604GB/s
ValidateLargeNonAscii         64243 ns      64234 ns      10771   1.45132GB/s
```
The new algorithm:
```
----------------------------------------------------------------
Benchmark                         Time           CPU Iterations
----------------------------------------------------------------
ValidateTinyAscii                 4 ns          4 ns  137163202   2.29073GB/s
ValidateTinyNonAscii             10 ns         10 ns   71646804   1068.54MB/s
ValidateSmallAscii               46 ns         46 ns   16725701   2.80417GB/s
ValidateSmallAlmostAscii         63 ns         63 ns   11153069   2.17771GB/s
ValidateSmallNonAscii            55 ns         55 ns   12845765   2.25957GB/s
ValidateLargeAscii            28394 ns      28390 ns      24765   3.28084GB/s
ValidateLargeAlmostAscii      27530 ns      27526 ns      21613    3.3851GB/s
ValidateLargeNonAscii         27492 ns      27488 ns      24760   3.39145GB/s
```