@pitrou 2019-08-12
What's the point of including this in Arrow if it's not used in the codebase?