@wesm 2019-08-13
I had been thinking that this would be exposed in some way to be invoked easily on `BinaryArray`. 

If this patch is accepted we need to add something to LICENSE.txt