@guyuqi 2019-08-16
> This also doesn't have any unit tests

I'd like to add the unit test case for `ValidateNonAscii`.