@guyuqi 2019-08-28
Add `ValidateNonAscii` to unittest case:  `arrow-utility-test`.
Arm64 Unit tests results:
```
Test project /home/builder/arrow/cpp/bld
      Start  2: arrow-array-test
      Start 52: arrow-io-compressed-test
      Start 35: arrow-ipc-read-write-test
      Start 26: arrow-compute-aggregate-test
 1/55 Test #26: arrow-compute-aggregate-test .......   Passed    0.90 sec
      Start 63: arrow-logging-test
 2/55 Test #35: arrow-ipc-read-write-test ..........   Passed    1.51 sec
 3/55 Test #63: arrow-logging-test .................   Passed    0.43 sec
      Start 31: arrow-compute-filter-test
      Start 66: arrow-thread-pool-test
 4/55 Test #66: arrow-thread-pool-test .............   Passed    0.47 sec
 5/55 Test #31: arrow-compute-filter-test ..........   Passed    0.47 sec
 6/55 Test #52: arrow-io-compressed-test ...........   Passed    2.22 sec
      Start 65: arrow-task-group-test
      Start 28: arrow-compute-compare-test
      Start 64: arrow-rle-encoding-test
 7/55 Test #65: arrow-task-group-test ..............   Passed    0.28 sec
      Start 55: arrow-io-memory-test
 8/55 Test #64: arrow-rle-encoding-test ............   Passed    0.38 sec
 9/55 Test #28: arrow-compute-compare-test .........   Passed    0.38 sec
      Start 24: arrow-compute-hash-test
      Start 30: arrow-compute-take-test
10/55 Test #24: arrow-compute-hash-test ............   Passed    0.18 sec
11/55 Test #55: arrow-io-memory-test ...............   Passed    0.38 sec
      Start 53: arrow-io-file-test
      Start 44: arrow-csv-parser-test
12/55 Test #30: arrow-compute-take-test ............   Passed    0.28 sec
13/55 Test #44: arrow-csv-parser-test ..............   Passed    0.20 sec
      Start 61: arrow-compression-test
      Start 56: arrow-io-readahead-test
14/55 Test #53: arrow-io-file-test .................   Passed    0.31 sec
15/55 Test #61: arrow-compression-test .............   Passed    0.20 sec
      Start  4: arrow-extension_type-test
      Start 43: arrow-csv-converter-test
16/55 Test #56: arrow-io-readahead-test ............   Passed    0.21 sec
17/55 Test #43: arrow-csv-converter-test ...........   Passed    0.07 sec
18/55 Test  #4: arrow-extension_type-test ..........   Passed    0.18 sec
      Start  1: arrow-allocator-test
      Start  3: arrow-buffer-test
      Start  8: arrow-result-test
19/55 Test  #8: arrow-result-test ..................   Passed    0.08 sec
20/55 Test  #3: arrow-buffer-test ..................   Passed    0.08 sec
21/55 Test  #1: arrow-allocator-test ...............   Passed    0.18 sec
      Start 41: arrow-csv-chunker-test
      Start 36: arrow-ipc-json-simple-test
      Start 38: arrow-json-integration-test
22/55 Test #38: arrow-json-integration-test ........   Passed    0.08 sec
23/55 Test #36: arrow-ipc-json-simple-test .........   Passed    0.08 sec
24/55 Test #41: arrow-csv-chunker-test .............   Passed    0.18 sec
      Start 18: arrow-compute-test
      Start 25: arrow-compute-util-internal-test
      Start 34: arrow-feather-test
25/55 Test #34: arrow-feather-test .................   Passed    0.09 sec
26/55 Test #25: arrow-compute-util-internal-test ...   Passed    0.09 sec
27/55 Test #18: arrow-compute-test .................   Passed    0.19 sec
      Start 22: arrow-compute-boolean-test
      Start 60: arrow-bit-util-test
      Start  7: arrow-public-api-test
28/55 Test  #7: arrow-public-api-test ..............   Passed    0.08 sec
29/55 Test #60: arrow-bit-util-test ................   Passed    0.08 sec
30/55 Test #22: arrow-compute-boolean-test .........   Passed    0.18 sec
      Start 11: arrow-stl-test
      Start 49: arrow-json-test
      Start 10: arrow-status-test
31/55 Test #10: arrow-status-test ..................   Passed    0.08 sec
32/55 Test #49: arrow-json-test ....................   Passed    0.10 sec
33/55 Test #11: arrow-stl-test .....................   Passed    0.20 sec
      Start 20: arrow-compute-operations-test
      Start 48: arrow-localfs-test
      Start 59: arrow-utility-test
34/55 Test #48: arrow-localfs-test .................   Passed    0.10 sec
35/55 Test #20: arrow-compute-operations-test ......   Passed    0.20 sec
      Start 13: arrow-table-test
      Start 16: arrow-sparse_tensor-test
36/55 Test #59: arrow-utility-test .................   Passed    0.21 sec
37/55 Test #16: arrow-sparse_tensor-test ...........   Passed    0.08 sec
38/55 Test #13: arrow-table-test ...................   Passed    0.18 sec
      Start 14: arrow-table_builder-test
      Start  6: arrow-pretty_print-test
      Start 47: arrow-filesystem-test
39/55 Test #47: arrow-filesystem-test ..............   Passed    0.08 sec
40/55 Test  #6: arrow-pretty_print-test ............   Passed    0.08 sec
41/55 Test #14: arrow-table_builder-test ...........   Passed    0.18 sec
      Start 62: arrow-decimal-test
      Start 78: arrow-dataset-file_test
      Start  5: arrow-memory_pool-test
42/55 Test  #5: arrow-memory_pool-test .............   Passed    0.08 sec
43/55 Test #78: arrow-dataset-file_test ............   Passed    0.08 sec
44/55 Test #62: arrow-decimal-test .................   Passed    0.18 sec
      Start 23: arrow-compute-cast-test
      Start 42: arrow-csv-column-builder-test
      Start 37: arrow-ipc-json-test
45/55 Test #37: arrow-ipc-json-test ................   Passed    0.08 sec
46/55 Test #42: arrow-csv-column-builder-test ......   Passed    0.09 sec
47/55 Test #23: arrow-compute-cast-test ............   Passed    0.19 sec
      Start 54: arrow-io-hdfs-test
      Start 40: arrow-concatenate-test
      Start 12: arrow-type-test
48/55 Test #12: arrow-type-test ....................   Passed    0.08 sec
49/55 Test #40: arrow-concatenate-test .............   Passed    0.08 sec
50/55 Test #54: arrow-io-hdfs-test .................   Passed    0.18 sec
      Start 51: arrow-io-buffered-test
      Start  9: arrow-scalar-test
      Start 15: arrow-tensor-test
51/55 Test #15: arrow-tensor-test ..................   Passed    0.08 sec
52/55 Test  #9: arrow-scalar-test ..................   Passed    0.08 sec
53/55 Test #51: arrow-io-buffered-test .............   Passed    0.19 sec
      Start 19: arrow-compute-expression-test
54/55 Test #19: arrow-compute-expression-test ......   Passed    0.10 sec
55/55 Test  #2: arrow-array-test ...................   Passed    6.17 sec

100% tests passed, 0 tests failed out of 55

Label Time Summary:
arrow-tests      =  19.49 sec*proc (54 tests)
arrow_dataset    =   0.08 sec*proc (1 test)
unittest         =  19.57 sec*proc (55 tests)

Total Test time (real) =   6.18 sec
[100%] Built target unittest

```