@wesm 2019-09-18
I'm working on adding a more capable Aarch64 machine to the Ursabot cluster (48-core ThunderX) so that could help with providing more timely CI