@pitrou 2019-09-18
Cool. But I still don't think we should hamper the full-ASCII case (which will be very common in e.g. CSV or JSON files).