@guyuqi 2019-09-19
> Cool. But I still don't think we should hamper the full-ASCII case (which will be very common in e.g. CSV or JSON files).

I'd like to propose a full-ASCII optimization for Arm64 later.