@guyuqi 2019-10-11
Add Ascii validation:
Before:
```
ValidateTinyAscii 6.04 ns 6.03 ns 112754472 bytes_per_second=1.54331G/s
ValidateSmallAscii 22.8 ns 22.7 ns 29583112 bytes_per_second=5.60858G/s
ValidateLargeAscii 15166 ns 15164 ns 46164 bytes_per_second=6.14226G/s

```
After:
```
ValidateTinyAscii 8.06 ns 8.06 ns 87046764 bytes_per_second=1.15497G/s
ValidateSmallAscii 14.5 ns 14.5 ns 48115118 bytes_per_second=8.77193G/s
ValidateLargeAscii 7840 ns 7839 ns 89218 bytes_per_second=11.8821G/s
```

Get ~ 1.9x speedup for Large Ascii